chore(build): comment out zip packaging functionality
Temporarily disables the creation of a zip archive as part of the build process by commenting out related code. This could be useful if the distribution packaging is not currently needed or if there's an intention to modify the current packaging workflow. Be cautious that this change means the final output will not include a zipped version of the build, potentially affecting deployment processes that rely on it.